requirements.txt
# Core Framework
streamlit==1.30.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# HuggingFace Client
huggingface-hub>=0.20.0

# Image Processing
Pillow==10.2.0

# File Upload
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Testing (optional for MVP)
pytest==7.4.3
pytest-asyncio==0.21.1
